PrepSprites/requirements.txt
pygame>=2.0.0
Pillow>=9.0.0
numpy>=1.20.0
